+++ mathics/builtin/drawing/graphics_internals.py
@@ -13,7 +13,7 @@ from mathics.builtin.base import (
 # Signals to Mathics doc processing not to include this module in its documentation.
 no_doc = True
 
-from mathics.core.symbols import system_symbols_dict
+from mathics.core.symbols import system_symbols_dict, Symbol
 
 
 class _GraphicsElement(InstanceableBuiltin):
@@ -30,10 +30,14 @@ class _GraphicsElement(InstanceableBuiltin):
         return klass(graphics, item)
 
 
-def get_class(name):
-    c = GLOBALS.get(name)
+def get_class(symbol: Symbol):
+    """
+    Returns the Builtin graphic primitive associated to the
+    Symbol `symbol`
+    """
+    c = GLOBALS.get(symbol)
     if c is None:
-        return GLOBALS3D.get(name)
+        return GLOBALS3D.get(symbol)
     else:
         return c